Send Suggestions to Compose Line
Created SuggestionModeSender Class to handle interacting with the InputLogic. First, the suggested text is selected, then the compose line is cleared before the suggested text is committed to the compose line. Private status was changed to public and static was added to a few variables in other classes to make this possible.